requirements.txt
google.cloud
os